reintroduce #103 issue
